move SDK-based functionality into separate file which isn't built for WASM
