+++ lint/lint.go
@@ -19,25 +19,18 @@
 package lint
 
 import (
-	"context"
 	"encoding/csv"
 	"fmt"
 	"io"
 	"log"
 	"os"
 	"path"
-	"sort"
 	"strings"
 	"sync"
-
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/onflow/cadence/common"
 	"github.com/onflow/cadence/pretty"
 	"github.com/onflow/cadence/tools/analysis"
-	"github.com/onflow/flow-go-sdk"
-	grpcAccess "github.com/onflow/flow-go-sdk/access/grpc"
 )
 
 const LoadMode = analysis.NeedTypes | analysis.NeedExtendedElaboration | analysis.NeedPositionInfo
@@ -74,124 +67,6 @@ func (l *Linter) PrettyPrintError(err error, location common.Location) {
 	if printErr != nil {
 		panic(printErr)
 	}
-}
-
-func (l *Linter) AnalyzeAccount(address string, networkName string) {
-	access, err := newFlowAccess(networkName)
-	if err != nil {
-		panic(err)
-	}
-
-	contractNames := map[common.Address][]string{}
-
-	getContracts := func(flowAddress flow.Address) (map[string][]byte, error) {
-		account, err := access.GetAccount(context.Background(), flowAddress)
-		if err != nil {
-			return nil, err
-		}
-
-		return account.Contracts, nil
-	}
-
-	flowAddress := flow.HexToAddress(address)
-	commonAddress := common.Address(flowAddress)
-
-	contracts, err := getContracts(flowAddress)
-	if err != nil {
-		panic(err)
-	}
-
-	locations := make([]common.Location, 0, len(contracts))
-	for contractName := range contracts {
-		location := common.AddressLocation{
-			Address: commonAddress,
-			Name:    contractName,
-		}
-		locations = append(locations, location)
-	}
-
-	analysisConfig := analysis.NewSimpleConfig(
-		LoadMode,
-		l.Codes,
-		contractNames,
-		func(address common.Address) (map[string][]byte, error) {
-			return getContracts(flow.Address(address))
-		},
-	)
-
-	l.analyze(analysisConfig, locations)
-}
-
-func (l *Linter) AnalyzeTransaction(transactionID flow.Identifier, networkName string) {
-	access, err := newFlowAccess(networkName)
-	if err != nil {
-		panic(err)
-	}
-
-	contractNames := map[common.Address][]string{}
-
-	getContracts := func(flowAddress flow.Address) (map[string][]byte, error) {
-		account, err := access.GetAccount(context.Background(), flowAddress)
-		if err != nil {
-			return nil, err
-		}
-
-		return account.Contracts, nil
-	}
-
-	transactionLocation := common.TransactionLocation(transactionID)
-
-	locations := []common.Location{
-		transactionLocation,
-	}
-
-	transaction, err := access.GetTransaction(context.Background(), transactionID)
-	if err != nil {
-		panic(err)
-	}
-
-	l.Codes[transactionLocation] = transaction.Script
-
-	analysisConfig := analysis.NewSimpleConfig(
-		LoadMode,
-		l.Codes,
-		contractNames,
-		func(address common.Address) (map[string][]byte, error) {
-			return getContracts(flow.Address(address))
-		},
-	)
-	l.analyze(analysisConfig, locations)
-}
-
-func newFlowAccess(networkName string) (*grpcAccess.Client, error) {
-	networkMap := map[string]string{
-		"mainnet":  grpcAccess.MainnetHost,
-		"testnet":  grpcAccess.TestnetHost,
-		"emulator": grpcAccess.EmulatorHost,
-		"":         grpcAccess.EmulatorHost,
-	}
-
-	network := networkMap[networkName]
-	if network == "" {
-		var names []string
-		for name := range networkMap {
-			names = append(names, name)
-		}
-		sort.Strings(names)
-
-		return nil, fmt.Errorf(
-			"missing network name. expected one of: %s",
-			strings.Join(names, ","),
-		)
-	}
-
-	return grpcAccess.NewClient(
-		network,
-		grpcAccess.WithGRPCDialOptions(
-			grpc.WithTransportCredentials(insecure.NewCredentials()),
-			grpc.WithDefaultCallOptions(),
-		),
-	)
 }
 
 func (l *Linter) AnalyzeCSV(path string) {
